fix: ensure EAS builds pull all Git LFS assets
Revert lazy-loading workarounds in the member area. The crash was caused
by wallet SVGs remaining as Git LFS pointers in EAS builds, which made
Metro fail to load the member route.

Scan all binary assets under src/assets for LFS pointers instead of
checking a fixed allowlist, pull when needed, and fail the build with a
clear error if assets are still missing.
