--- src/components/Member/wallet-pass-button.web.tsx
@@ -1,11 +1,0 @@
-import type React from 'react'
-
-interface WalletPassButtonProps {
-	onComplete: () => void
-}
-
-export function WalletPassButton({
-	onComplete: _onComplete
-}: WalletPassButtonProps): React.JSX.Element | null {
-	return null
-}
